requirements.txt
numpy==2.2.6
opencv-python-headless==4.11.0.86
pupil-apriltags==1.0.4.post11
pymodbus==3.9.2
PyYAML==6.0.2